+++ pdstable/__init__.py
@@ -52,6 +52,9 @@ import numbers
 import pdsparser
 import julian
 
+from .pds4table import Pds4TableInfo
+from .pds3table import Pds3TableInfo
+
 try:
     from ._version import __version__
 except ImportError as err:
@@ -79,6 +82,7 @@ FILE_SPECIFICATION_COLUMN_NAMES = (
     'FILE_SPECIFICATION_NAME',
     'FILE SPECIFICATION NAME',
     'FILE_NAME',
+    'File Name'
     'FILE NAME',
     'FILENAME',
     'PRODUCT_ID',
@@ -187,11 +191,19 @@ class PdsTable(object):
 
         # Parse the label
         if label_contents is not None:
-            self.info = PdsTableInfo(label_file, label_list=label_contents,
-                                     invalid=invalid, valid_ranges=valid_ranges)
+            if label_file.endswith(".xml"):
+                self.info = Pds4TableInfo(label_file, label_list=label_contents,
+                                        invalid=invalid, valid_ranges=valid_ranges)
+            else:
+                self.info = Pds3TableInfo(label_file, label_list=label_contents,
+                                        invalid=invalid, valid_ranges=valid_ranges)
         else:
-            self.info = PdsTableInfo(label_file,
-                                     invalid=invalid, valid_ranges=valid_ranges)
+            if label_file.endswith(".xml"):
+                self.info = Pds4TableInfo(label_file,
+                                        invalid=invalid, valid_ranges=valid_ranges)
+            else:
+                self.info = Pds3TableInfo(label_file,
+                                        invalid=invalid, valid_ranges=valid_ranges)
 
         # Select the columns
         if len(columns) == 0:
@@ -212,6 +224,10 @@ class PdsTable(object):
                 lines = f.readlines()
 
             # Check line count
+            # In PDS4, skip the first line because it's the column names
+            if label_file.endswith(".xml"):
+                lines = lines[1:]
+
             if len(lines) != self.info.rows:
                 raise ValueError('row count mismatch in %s: ' % label_file +
                                  '%d rows in file; ' % len(lines) +
@@ -221,12 +237,16 @@ class PdsTable(object):
             self.first = row_range[0]
             self.rows = row_range[1] - row_range[0]
 
-            record_bytes = self.info.label['RECORD_BYTES'].value
+            if label_file.endswith(".xml"):
+                # record_bytes is stored in row_bytes for PDS4
+                record_bytes = self.info.row_bytes
+            else:
+                record_bytes = self.info.label['RECORD_BYTES'].value
+
             with open(self.info.table_file_path, "rb") as f:
                 f.seek(row_range[0] * record_bytes)
                 lines = f.readlines(self.rows * record_bytes - 1)
 
-            # Check line count
             if len(lines) > self.rows:
                 lines = lines[:self.rows]
 
@@ -976,189 +996,3 @@ def lowercase_value(value):
         value_lc = value
 
     return value_lc
-
-################################################################################
-# Class PdsTableInfo
-################################################################################
-
-class PdsTableInfo(object):
-    """The PdsTableInfo class holds the attributes of a PDS-labeled table."""
-
-    def __init__(self, label_file_path, label_list=None, invalid={},
-                                                         valid_ranges={}):
-        """Loads a PDS table based on its associated label file.
-
-        Input:
-            label_file_path path to the label file
-            label_list      an option to override the parsing of the label.
-                            If this is a list, it is interpreted as containing
-                            all the records of the PDS label, in which case the
-                            overrides the contents of the label file.
-                            Alternatively, this can be a PdsLabel object that
-                            was already parsed.
-            invalid         an optional dictionary keyed by column name. The
-                            returned value must be a list or set of values that
-                            are to be treated as invalid, missing or unknown.
-            valid_ranges    an optional dictionary keyed by column name. The
-                            returned value must be a tuple or list containing
-                            the minimum and maximum numeric values in that
-                            column.
-        """
-
-        # Parse the label
-        if label_list is None:
-            self.label = pdsparser.PdsLabel.from_file(label_file_path)
-        elif isinstance(label_list, pdsparser.PdsLabel):
-            self.label = label_list
-        else:
-            self.label = pdsparser.PdsLabel.from_string(label_list)
-
-        # Get the basic file info...
-        if self.label["RECORD_TYPE"].value != "FIXED_LENGTH":
-            raise IOError('PDS table does not contain fixed-length records')
-
-        # Find the pointer to the table file
-        # Confirm that the value is a PdsSimplePointer
-        self.table_file_name = None
-        for node in self.label:
-            if node.name[0] == "^":
-                pointer_name = node.name[1:]
-                if isinstance(node.pdsvalue, pdsparser.PdsOffsetPointer):
-                    self.table_file_name = node.pdsvalue.value
-                    msg = ("Table file pointer '" + str(node.pdsvalue) +
-                           " is not a Simple Pointer and isn't fully "+
-                           "supported")
-                    warnings.warn(msg)
-                else:
-                    assert isinstance(node.pdsvalue, pdsparser.PdsSimplePointer)
-                    self.table_file_name = node.pdsvalue.value
-
-        if self.table_file_name is None:
-            raise IOError("Pointer to a data file was not found in PDS label")
-
-        # Locate the root of the table object
-        table_node = self.label[pointer_name]
-
-        # Save key info about the table
-        if table_node["INTERCHANGE_FORMAT"].value != "ASCII":
-            raise IOError('PDS table is not in ASCII format')
-
-        self.rows = table_node["ROWS"].value
-        self.columns = table_node["COLUMNS"].value
-        self.row_bytes = table_node["ROW_BYTES"].value
-
-        # Save the key info about each column in a list and a dictionary
-        self.column_info_list = []
-        self.column_info_dict = {}
-
-        # Construct the dtype0 dictionary
-        self.dtype0 = {'crlf': ('|S2', self.row_bytes-2)}
-
-        default_invalid = set(invalid.get("default", []))
-        counter = 0
-        for node in table_node:
-            if node.pdsvalue.value == "COLUMN":
-                node_dict = node.as_python_value()
-                name = node_dict["NAME"]
-                pdscol = PdsColumnInfo(node_dict, counter,
-                            invalid = invalid.get(name, default_invalid),
-                            valid_range = valid_ranges.get(name, None))
-                counter += 1
-
-                if name in self.column_info_dict:
-                    raise ValueError('duplicated column name: ' + name)
-
-                self.column_info_list.append(pdscol)
-                self.column_info_dict[pdscol.name] = pdscol
-                self.dtype0[pdscol.name] = pdscol.dtype0
-
-        # Fill in the complete table file name
-        self.table_file_path = os.path.join(os.path.dirname(label_file_path),
-                                            self.table_file_name)
-
-################################################################################
-# class PdsColumnInfo
-################################################################################
-
-class PdsColumnInfo(object):
-    """The PdsColumnInfo class holds the attributes of one column in a PDS
-    label."""
-
-    def __init__(self, node_dict, column_no, invalid=set(), valid_range=None):
-        """Constructor for a PdsColumn.
-
-        Input:
-            node_dict   the dictionary associated with the pdsparser.PdsNode
-                        object defining the column.
-            column_no   the index number of this column, starting at zero.
-            invalid     an optional set of discrete values that are to be
-                        treated as invalid, missing or unknown.
-            valid_range an optional tuple or list identifying the lower and
-                        upper limits of the valid range for a numeric column.
-        """
-
-        self.name = node_dict["NAME"]
-        self.colno = column_no
-
-        self.start_byte = node_dict["START_BYTE"]
-        self.bytes      = node_dict["BYTES"]
-
-        self.items = node_dict.get("ITEMS", 1)
-        self.item_bytes = node_dict.get("ITEM_BYTES", self.bytes)
-        self.item_offset = node_dict.get("ITEM_OFFSET", self.bytes)
-
-        # Define dtype0 to isolate each column in a record
-        self.dtype0 = ("S" + str(self.bytes), self.start_byte - 1)
-
-        # Define dtype1 as a list of dtypes needed to isolate each item
-        if self.items == 1:
-            self.dtype1 = None
-        else:
-            self.dtype1 = {}
-            byte0 = 0
-            for i in range(self.items):
-                self.dtype1["item_" + str(i)] = ("S" + str(self.item_bytes),
-                                                 byte0)
-                byte0 += self.item_offset
-
-        # Define dtype2 as the intended dtype of the values in the column
-        self.data_type = node_dict["DATA_TYPE"]
-        if "INTEGER" in self.data_type:
-            self.data_type = "int"
-            self.dtype2 = "int"
-            self.scalar_func = int
-        elif "REAL" in self.data_type:
-            self.data_type = "float"
-            self.dtype2 = "float"
-            self.scalar_func = float
-        elif ("TIME" in self.data_type or "DATE" in self.data_type or
-              self.name.endswith("_TIME") or self.name.endswith("_DATE")):
-            self.data_type = "time"
-            self.dtype2 = 'S'
-            self.scalar_func = tai_from_iso
-        elif "CHAR" in self.data_type:
-            self.data_type = "string"
-            self.dtype2 = STR_DTYPE
-            self.scalar_func = None
-        else:
-            raise IOError("unsupported data type: " + data_type)
-
-        # Identify validity criteria
-        self.valid_range = valid_range or node_dict.get("VALID_RANGE", None)
-
-        if isinstance(invalid, (numbers.Real,) + STRING_TYPES):
-            invalid = set([invalid])
-
-        self.invalid_values = set(invalid)
-
-        self.invalid_values.add(node_dict.get("INVALID_CONSTANT"       , None))
-        self.invalid_values.add(node_dict.get("MISSING_CONSTANT"       , None))
-        self.invalid_values.add(node_dict.get("UNKNOWN_CONSTANT"       , None))
-        self.invalid_values.add(node_dict.get("NOT_APPLICABLE_CONSTANT", None))
-        self.invalid_values.add(node_dict.get("NULL_CONSTANT"          , None))
-        self.invalid_values.add(node_dict.get("INVALID"                , None))
-        self.invalid_values.add(node_dict.get("MISSING"                , None))
-        self.invalid_values.add(node_dict.get("UNKNOWN"                , None))
-        self.invalid_values.add(node_dict.get("NOT_APPLICABLE"         , None))
-        self.invalid_values.add(node_dict.get("NULL"                   , None))
-        self.invalid_values -= {None}
